# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MDG
  version: '1'
  summary: Self-hosted personal accounting service.
  description: |
    MDG ("Moi.Den.Gi") is a self-hosted personal accounting web service. This
    document describes API v1 and is the only specification of that API — it is
    hand-written and hand-maintained, and nothing in the build verifies it
    against the code. If you change a controller, change this too.

    ## Content type

    Every request and response uses the media type
    `application/vnd.mdg+json;version=1` — errors included, and with no
    `charset` parameter. Write operations require it as a `Content-Type` request
    header, even the two maintenance triggers that take no body.

    The only responses without it are the `204 No Content` ones.

    ## Errors

    Errors are RFC 7807-shaped but carry the vendor media type rather than
    `application/problem+json`, and add a machine-readable `code` field.

    ## Authentication

    There is none. 
    
    ## Paging

    `GET /transactions` and `GET /budgets` are cursor-paged. Pass `limit` to
    start paging, then follow the `next` cursor. Cursors are **self-contained** —
    they carry the original filter, sort, embed and limit — so when `cursor` is
    supplied every other query parameter is ignored, though a `limit` below one
    is still rejected rather than quietly dropped. `first` and `next` come
    back as the empty string when there is no such page, and handing an empty
...

    ## Embedding

...

...
servers:
...
paths:
...
  /budgets:
    get:
      tags: [Budgets]
      summary: List all budgets
      description: |
        Ordered by `term_beginning` descending.

        Only `id`, `term_beginning` and `term_end` are populated here — the
        computed fields (`incoming_amount`, `outgoing_amount`, `state`) are
        omitted from the collection and must be read from a single budget.
      operationId: listBudgets
      parameters:
        - $ref: './components/parameters.yaml#/Limit'
        - $ref: './components/parameters.yaml#/Cursor'
      responses:
        '200':
          description: One page of budgets.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/budget.yaml#/Budgets'
          links:
            GetListedBudget:
              operationId: getBudget
              description: Read the first budget of this page, with its computed totals.
              parameters:
                path.id: '$response.body#/budgets/0/id'
            UpdateListedBudget:
              operationId: updateBudget
              description: Submit the first budget's own term back to it.
              parameters:
                path.id: '$response.body#/budgets/0/id'
              requestBody:
                term_beginning: '$response.body#/budgets/0/term_beginning'
                term_end: '$response.body#/budgets/0/term_end'
              x-schemathesis:
                merge_body: false
            ListListedBudgetEntries:
              operationId: listBudgetEntries
              description: List the entries of the first budget of this page.
              parameters:
                path.budgetId: '$response.body#/budgets/0/id'
            CopyListedBudgetEntries:
              operationId: copyBudgetEntries
              description: |
                Copy the first budget's entry amounts onto itself — the one source that
                is guaranteed to exist for any budget in the page.
              parameters:
                path.budgetId: '$response.body#/budgets/0/id'
                path.sourceBudgetId: '$response.body#/budgets/0/id'
                path.mode: preserve
            DeleteListedBudget:
              operationId: deleteBudget
              description: Delete the first budget of this page and all of its entries.
              parameters:
                path.id: '$response.body#/budgets/0/id'
            NextBudgetPage:
              operationId: listBudgets
              description: |
                Follow this page's `next` cursor. An exhausted page carries the empty
                string, which reads as *no cursor* and returns the first page again.
              parameters:
                query.cursor: '$response.body#/next'
        '400':
          $ref: './components/responses.yaml#/BadRequest'
        '422':
          $ref: './components/responses.yaml#/UnprocessableEntity'
    post:
      tags: [Budgets]
      summary: Create a new budget
      description: |
        Only `term_beginning` and `term_end` are read; every other field is
        computed. Terms are **inclusive** at both ends.

        Three rules apply, each with the `code` it fails under:

        * `term_end` must not precede `term_beginning` — `BUDGET_INVALID_TERM`;
        * the term must span at least two days — `BUDGET_SHORT_RANGE`;
        * terms of different budgets must not overlap — `BUDGET_OVERLAPPING`.

        Creating a budget also creates one budget entry per income and expense
        account. The budget's `id` is derived from `term_beginning` as `YYYYMMDD`.
      operationId: createBudget
      requestBody:
        required: true
        content:
          application/vnd.mdg+json;version=1:
            schema:
              $ref: './components/schemas/budget.yaml#/Budget'
            example:
              term_beginning: '2017-02-04'
              term_end: '2017-02-26'
      responses:
        '201':
          description: The created budget.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/budget.yaml#/Budget'
          links:
            GetBudget:
              operationId: getBudget
              description: Read the new budget together with its computed totals.
              parameters:
                path.id: '$response.body#/id'
            UpdateBudget:
              operationId: updateBudget
              description: |
                Submit the budget's own term back to it. The overlap check ignores the
                budget being updated, so an unchanged term is accepted rather than
                failing with `BUDGET_OVERLAPPING`.
              parameters:
                path.id: '$response.body#/id'
              requestBody:
                term_beginning: '$response.body#/term_beginning'
                term_end: '$response.body#/term_end'
              x-schemathesis:
                merge_body: false
            ListBudgetEntries:
              operationId: listBudgetEntries
              description: |
                List the entries created together with the budget — one per income and
                expense account that existed when it was created.
              parameters:
                path.budgetId: '$response.body#/id'
            GetBudgetEntryTree:
              operationId: getBudgetEntryTree
              description: The same entries, arranged into the category tree.
              parameters:
                path.budgetId: '$response.body#/id'
            CopyBudgetEntries:
              operationId: copyBudgetEntries
              description: |
                Copy entry amounts into the new budget. Source and target are the same
                budget here, that being the only source guaranteed to exist at this
                point; both ids are resolved exactly, so the call answers 204.
              parameters:
                path.budgetId: '$response.body#/id'
                path.sourceBudgetId: '$response.body#/id'
                path.mode: preserve
            DeleteBudget:
              operationId: deleteBudget
              description: Delete the new budget and every entry belonging to it.
              parameters:
                path.id: '$response.body#/id'
        '400':
          $ref: './components/responses.yaml#/BadRequest'
        '412':
          $ref: './components/responses.yaml#/PreconditionFailed'
        '415':
          $ref: './components/responses.yaml#/UnsupportedMediaType'
        '422':
          $ref: './components/responses.yaml#/UnprocessableEntity'
  /budgets/{id}:
    get:
      tags: [Budgets]
      summary: View a budget
      description: |
        The path segment may be a budget id or **any date inside a budget's term**
        in `YYYYMMDD` form; the budget whose `term_beginning` is closest to and
        not after that date is returned.

        Unlike the collection, this response carries the computed
        `incoming_amount`, `outgoing_amount` and `state`.
      operationId: getBudget
      parameters:
        - $ref: './components/parameters.yaml#/BudgetIdPath'
      responses:
        '200':
          description: The budget, with its computed totals.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/budget.yaml#/Budget'
        '404':
          description: No budget found
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/problem.yaml#/Problem'
    put:
      tags: [Budgets]
      summary: Modify a budget
      description: |
        Changes the budget's term.

        The path segment resolves exactly as it does on `GET` and `DELETE`: it
        may be a budget id or **any date inside a budget's term** in `YYYYMMDD`
        form, and the budget whose `term_beginning` is closest to and not after
        that date is the one updated.

        The same three term rules as on create apply, except that the overlap
        check ignores the budget being updated — a term may be kept as-is,
        shrunk, or extended into free space without tripping
        `BUDGET_OVERLAPPING`.

        The budget's `id` is **not** recomputed from a new `term_beginning`; it
        keeps the id it was created with.
      operationId: updateBudget
      parameters:
        - $ref: './components/parameters.yaml#/BudgetIdPath'
      requestBody:
        required: true
        content:
          application/vnd.mdg+json;version=1:
            schema:
              $ref: './components/schemas/budget.yaml#/Budget'
            example:
              term_beginning: '2017-02-04'
              term_end: '2017-02-28'
      responses:
        '202':
          description: The updated budget.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/budget.yaml#/Budget'
        '400':
          $ref: './components/responses.yaml#/BadRequest'
        '404':
          description: No budget found
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/problem.yaml#/Problem'
        '412':
          $ref: './components/responses.yaml#/PreconditionFailed'
        '415':
          $ref: './components/responses.yaml#/UnsupportedMediaType'
        '422':
          $ref: './components/responses.yaml#/UnprocessableEntity'
    delete:
      tags: [Budgets]
      summary: Delete a budget
      description: |
        All of the budget's entries are deleted with it.

        The path segment resolves exactly as it does on `GET`: it may be a budget
        id or **any date inside a budget's term** in `YYYYMMDD` form, and the
        budget whose `term_beginning` is closest to and not after that date is the
        one deleted. `DELETE /budgets/20170205` therefore deletes the budget that
        began on `2017-02-04`.
      operationId: deleteBudget
      parameters:
        - $ref: './components/parameters.yaml#/BudgetIdPath'
      responses:
        '204':
          description: The budget was deleted.
        '404':
          description: No budget found
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/problem.yaml#/Problem'
  /budgets/{budgetId}/entries:
    get:
      tags: [Budget entries]
      summary: List budget entries
      description: |
        Every entry of the budget, flat. Entries exist for income and expense
        accounts only — never for asset accounts.
      operationId: listBudgetEntries
      parameters:
        - $ref: './components/parameters.yaml#/OwningBudgetId'
        - $ref: './components/parameters.yaml#/EmbedAccount'
      responses:
        '200':
          description: The budget's entries.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/budget.yaml#/BudgetEntries'
          links:
            GetBudgetEntry:
              operationId: getBudgetEntry
              description: Read the first entry of this listing.
              parameters:
                path.budgetId: '$request.path.budgetId'
                path.entryId: '$response.body#/budget_entries/0/id'
            UpdateBudgetEntry:
              operationId: updateBudgetEntry
              description: |
                Put an expected amount on the first entry of this listing. Only
                `distribution`, `expected_amount` and `dt` are writable.
              parameters:
                path.budgetId: '$request.path.budgetId'
                path.entryId: '$response.body#/budget_entries/0/id'
              requestBody:
                distribution: EVEN
                expected_amount: 100
              x-schemathesis:
                merge_body: false
        '404':
          $ref: './components/responses.yaml#/NotFound'
  /budgets/{budgetId}/entries/tree:
    get:
      tags: [Budget entries]
      summary: Get budget entries as a category tree
      description: |
        The budget's entries arranged into the category forest, split into
        `expense` and `income`. Every node carries the totals of everything
        beneath it.

        A category node's distribution semantics are derived from its children: if
        any descendant entry is `SINGLE` the whole category is treated as
        `SINGLE`; failing that, if any is `EVEN` the category is `EVEN`; otherwise
        `PRORATED`.

        Accounts embedded here are not populated with their categories, and report
        `balance: 0` / `primary_balance: 0` like every other embedded account.
      operationId: getBudgetEntryTree
      parameters:
        - $ref: './components/parameters.yaml#/OwningBudgetId'
        - $ref: './components/parameters.yaml#/EmbedAccount'
        - $ref: './components/parameters.yaml#/BudgetEntryFilter'
      responses:
        '200':
          description: The budget entry tree.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/budget.yaml#/BudgetEntryTree'
        '404':
          $ref: './components/responses.yaml#/NotFound'
  /budgets/{budgetId}/entries/{entryId}:
    get:
      tags: [Budget entries]
      summary: View a budget entry
      operationId: getBudgetEntry
      parameters:
        - $ref: './components/parameters.yaml#/OwningBudgetId'
        - $ref: './components/parameters.yaml#/BudgetEntryId'
        - $ref: './components/parameters.yaml#/EmbedAccount'
      responses:
        '200':
          description: The budget entry.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/budget.yaml#/BudgetEntry'
        '404':
          description: No entry found
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/problem.yaml#/Problem'
    put:
      tags: [Budget entries]
      summary: Modify a budget entry
      description: |
        Only `distribution`, `expected_amount` and `dt` are applied — everything
        else is computed or immutable. A negative `expected_amount` is rejected
        with 422 `BUDGETENTRY_IS_NEGATIVE`, and a `dt` outside the budget's term
        with 409 `BUDGETENTRY_DT_OUT_OF_BUDGET`.

        `dt` only means anything for `SINGLE` distribution: it is silently dropped
        for `EVEN` and `PRORATED`, and switching an entry away from `SINGLE`
        removes a previously stored value.

      operationId: updateBudgetEntry
      parameters:
        - $ref: './components/parameters.yaml#/OwningBudgetId'
        - $ref: './components/parameters.yaml#/BudgetEntryId'
      requestBody:
        required: true
        content:
          application/vnd.mdg+json;version=1:
            schema:
              $ref: './components/schemas/budget.yaml#/BudgetEntry'
            example:
              id: 291
              account_id: 2065
              distribution: SINGLE
              expected_amount: 1000
              dt: '2017-02-15'
      responses:
        '202':
          description: The updated budget entry.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/budget.yaml#/BudgetEntry'
        '400':
          $ref: './components/responses.yaml#/BadRequest'
        '404':
          description: No entry found
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/problem.yaml#/Problem'
        '409':
          $ref: './components/responses.yaml#/Conflict'
        '415':
          $ref: './components/responses.yaml#/UnsupportedMediaType'
        '422':
          $ref: './components/responses.yaml#/UnprocessableEntity'
  /budgets/{budgetId}/entries/copy/{mode}/{sourceBudgetId}:
    put:
      tags: [Budget entries]
      summary: Copy budget entries from another budget
      description: |
        Copies `expected_amount` values from the source budget's entries into this
        budget's entries, matching on account.

        Takes no request body and returns no content. A `Content-Type`
        header is not required here, unlike on the other write operations.

        Entries that are set in the target but zero in the source are left alone
        in both modes.
      operationId: copyBudgetEntries
      parameters:
        - $ref: './components/parameters.yaml#/OwningBudgetId'
        - name: mode
          in: path
          required: true
          description: |
            `overwrite` replaces the target's values; **any other value**, `preserve`
            by convention, keeps values the target has already customised. The
            comparison is a case-insensitive equality test against the literal
            `overwrite`, so an unrecognised mode behaves as `preserve` rather than
            erroring.
          schema:
            type: string
            enum: [preserve, overwrite]
          example: preserve
        - name: sourceBudgetId
          in: path
          required: true
          description: Id of the budget to copy from, as `YYYYMMDD`.
          schema:
            type: integer
            format: int64
          example: 20170101
      responses:
        '204':
          description: The entries were copied.
        '404':
          description: Source budget could not be found
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/problem.yaml#/Problem'
...